+++ provider/Cargo.toml
@@ -5,18 +5,17 @@ authors = ["James Prestwich <james@prestwi.ch>"]
 edition = "2018"
 
 [dependencies]
+# core utilities
 async-trait = "0.1.36"
-
 futures-core = { version = "0.3.5", default-features = false }
-futures-util = { version = "0.3.5", default-features = false, features = ["std"] }
+futures-util = { version = "0.3.5", default-features = false, features = ["std", "channel"] }
 futures-timer = "3.0.2"
 pin-project = { version = "0.4.20", default-features = false }
+lru = { version = "0.5.2" }
 
+# other projects in this workspace
 riemann-core = { path = "../core" }
 rmn-btc = { path = "../bitcoin", default-features = false }
-
-# esplora cache only
-lru = { version = "0.5.2" }
 
 # RPC only
 secrecy = { version = "0.7.0", optional = true }
@@ -28,6 +27,7 @@ hex = { version = "0.4.2", optional = true }
 serde = { version = "1.0.106", features = ["derive"], optional = true }
 serde_json = { version = "1.0.55", optional = true }
 bytes = { version = "^0.5", optional = true }
+
 # building wasm
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen = { version = "0.2.62", features = ["serde-serialize"]  }
